template - move <Icons> into /ui/common
