+++ code/compiler/template.esm/src/ui/common/Icons.ts
